+++ src/basics/mod.rs
@@ -10,7 +10,7 @@ pub use self::config::{EvalConfig, ExecutionMode, Statistics, TimeFormat, TimeRe
 pub use self::io_handler::OutputChannel;
 pub(crate) use self::io_handler::{create_event_source, EventSource, EventSourceConfig, OutputHandler};
 
-pub use self::csv_input::{CSVEventSource, CSVInputSource};
+pub use self::csv_input::{CsvEventSource, CsvInputSource};
 
 #[cfg(feature = "pcap_interface")]
 pub use self::pcap_input::{PCAPEventSource, PCAPInputSource};