+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [0.6.1] - ?
 
 ### Added
 - `From` implementation for values
+- New window operations: last, median, nth-percentile, variance, standard deviation, covariance
 
 ## [0.6.0] - 2021-08-03